adding error handling
change private to true inside package.json to enable npm to publish it.
adding CORS proxy URL to accuWeatherFunctions.js functions
to prevent CORS error that occur when receiving an error when calling fetch to the accuweather API.
adding error handling when calling the accuWeatherFunctions.js functions using useState and presenting a Modal   from bootstrap presenting a message to the user.